backend/requirements.txt
google-generativeai
duckduckgo-search
networkx
pyvis
colorama
requests